@@ -11,20 +11,20 @@
         }
     ],
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-json": "*",
-        "illuminate/log": "^10.0",
-        "illuminate/support": "^10.0",
+        "illuminate/log": "^11.0",
+        "illuminate/support": "^11.0",
         "spatie/backtrace": "^1.2",
         "guzzlehttp/guzzle": "^7.5"
     },
     "require-dev": {
-        "orchestra/testbench": "v8.0",
-        "phpunit/phpunit": "^9.3.3",
         "phpmd/phpmd": "^2.9",
         "squizlabs/php_codesniffer": "^3.6",
         "friendsofphp/php-cs-fixer": "^3.0",
-        "brianium/paratest": "^6.2"
+        "brianium/paratest": "^7.4",
+        "phpunit/phpunit": "^10.5",
+        "orchestra/testbench": "^9.0"
     },
     "autoload": {
         "psr-4": {
