@@ -11,7 +11,7 @@ use Symfony\Component\VarDumper\Dumper\HtmlDumper as BaseHtmlDumper;
  */
 class HtmlDumper extends BaseHtmlDumper
 {
-    protected $dumpHeader = '';
+    protected ?string $dumpHeader = null;
 
     public function dumpVariable($variable): string
     {
